pyproject.toml
[tool.poetry]
name = "zeply-test"
version = "0.1.0"
description = ""
authors = ["Eugene Savchenko <savchenko.eo@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "~3.10"
Django = "^4.1.7"
djangorestframework = "^3.14.0"
python-decouple = "^3.8"
django-cryptography = "^1.1"
xrpl-py = "^1.7.0"
django-stubs = "^1.15.0"
bip32utils = "^0.3.post4"
pysha3 = "^1.0.2"
python-dotenv = "^1.0.0"
factory-boy = "^3.2.1"
gunicorn = "^20.1.0"
whitenoise = "^6.4.0"
psycopg2-binary = "^2.9.5"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
isort = "^5.12.0"
flake8 = "^6.0.0"
black = "^23.1.0"
pytest-django = "^4.5.2"
ipython = "^8.11.0"
pre-commit = "^3.1.1"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "zeplytest.settings"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
